Split error components
